requirements.txt
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
httplib2==0.20.2
jsonschema==3.2.0
PyJWT==2.3.0
PyMySQL==1.0.2
django-environ==0.10.0
whitenoise==6.5.0
pandas==2.1.1
cryptography==41.0.5